New option: TrackChanges
When activated, the current results are compared to the results in the output directory. Whenever the number of fragments or location changes, the result files from the output directory are copied to the ChangesJournal sub-directory. Additionally, a new entry will be added to the changes.json file to record the new number of fragments and locations.